+++ packages/adapters/src/adapters/Cornerstone3D/Probe.ts
@@ -30,8 +30,6 @@ class Probe extends BaseAdapter3D {
         const {
             state,
             NUMGroup,
-            scoord,
-            scoordArgs,
             worldCoords,
             referencedImageId,
             ReferencedFrameNumber
@@ -45,21 +43,20 @@ class Probe extends BaseAdapter3D {
         const cachedStats = referencedImageId
             ? {
                   [`imageId:${referencedImageId}`]: {
-                      length: NUMGroup?.MeasuredValueSequence?.NumericValue || 0
+                      value:
+                          NUMGroup?.MeasuredValueSequence?.NumericValue ?? null
                   }
               }
             : {};
         state.annotation.data = {
             ...state.annotation.data,
             handles: {
-                points: worldCoords,
-                activeHandleIndex: 0,
-                textBox: {
-                    hasMoved: false
-                }
+                ...state.annotation.data.handles,
+                points: worldCoords
             },
             cachedStats,
-            frameNumber: ReferencedFrameNumber
+            frameNumber: ReferencedFrameNumber,
+            invalidated: true
         };
 
         return state;