+++ packages/adapters/src/adapters/Cornerstone3D/UltrasoundDirectional.ts
@@ -27,11 +27,8 @@ class UltrasoundDirectional extends BaseAdapter3D {
         state.annotation.data = {
             ...state.annotation.data,
             handles: {
-                points: worldCoords,
-                activeHandleIndex: 0,
-                textBox: {
-                    hasMoved: false
-                }
+                ...state.annotation.data.handles,
+                points: worldCoords
             },
             frameNumber: ReferencedFrameNumber
         };