+++ packages/tools/src/tools/segmentation/CircleROIStartEndThresholdTool.ts
@@ -187,7 +187,7 @@ class CircleROIStartEndThresholdTool extends CircleROITool {
         Types.Point3,
         Types.Point3,
         Types.Point3,
-        Types.Point3
+        Types.Point3,
       ];
     }